requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
awslambdaric==2.0.8
certifi==2023.11.17
charset-normalizer==3.3.2
deprecation==2.1.0
exceptiongroup==1.2.0
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
outcome==1.3.0.post0
packaging==23.2
postgrest==0.13.2
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.2
requests==2.31.0
selenium==4.16.0
simplejson==3.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.3
supafunc==0.3.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
webdriver-manager==4.0.1
websockets==11.0.3
wsproto==1.2.0
